Made py tp external name changes
